+++ packages/connectors/hull-customerio/server/lib/sync-agent.js
@@ -22,6 +22,8 @@ const HashUtil = require("./sync-agent/hash-util");
 const ValidationUtil = require("./sync-agent/validation-util");
 const SHARED_MESSAGES = require("./shared-messages");
 
+const { getNumberFromContext } = require("./utils");
+
 const BASE_API_URL = "https://track.customer.io";
 const SEGMENT_PROPERTY_NAME = "segments"; // Prep for transition to dedicated segments for accounts and users
 
@@ -172,11 +174,27 @@ class SyncAgent {
     // Init the hash util
     this.hashUtil = new HashUtil();
 
+    const maxAttributeNameLength = getNumberFromContext(
+      reqContext,
+      "connector.private_settings.max_attribute_name_length",
+      150
+    );
+    const maxAttributeValueLength = getNumberFromContext(
+      reqContext,
+      "connector.private_settings.max_attribute_value_length",
+      1000
+    );
+    const maxIdentifierValueLength = getNumberFromContext(
+      reqContext,
+      "connector.private_settings.max_identifier_value_length",
+      150
+    );
+
     // Init the validation util
     const valUtilOptions: IValidationUtilOptions = {
-      maxAttributeNameLength: 150,
-      maxAttributeValueLength: 1000,
-      maxIdentifierValueLength: 150
+      maxAttributeNameLength,
+      maxAttributeValueLength,
+      maxIdentifierValueLength
     };
 
     this.validationUtil = new ValidationUtil(valUtilOptions);
@@ -345,14 +363,12 @@ class SyncAgent {
     return this.serviceClient
       .updateCustomer(envelope.customer)
       .then(() => {
-        return userScopedClient
-          .traits(userTraits, { source: "customerio" })
-          .then(() => {
-            userScopedClient.logger.info("outgoing.user.success", {
-              data: envelope.customer,
-              operation: "updateCustomer"
-            });
+        return userScopedClient.traits(userTraits).then(() => {
+          userScopedClient.logger.info("outgoing.user.success", {
+            data: envelope.customer,
+            operation: "updateCustomer"
           });
+        });
       })
       .then(() => {
         // Process the events
@@ -440,16 +456,13 @@ class SyncAgent {
       .deleteCustomer(_.get(envelope, "customer.id"))
       .then(() => {
         return userScopedClient
-          .traits(
-            {
-              deleted_at: new Date(),
-              id: null,
-              hash: null,
-              synced_at: null,
-              created_at: null
-            },
-            { source: "customerio" }
-          )
+          .traits({
+            "customerio/deleted_at": new Date(),
+            "customerio/id": null,
+            "customerio/hash": null,
+            "customerio/synced_at": null,
+            "customerio/created_at": null
+          })
           .then(() => {
             return userScopedClient.logger.info("outgoing.user.deletion", {
               data: { id: envelope.customer.id },