+++ packages/connectors/hull-customerio/test/integration/scenarios/customer-update-event-insert/smart-notifier-payload.js
@@ -7,9 +7,9 @@ module.exports = () => {
   const userSegmentId = _.get(smartNotifierPayload, "messages[0].segments[0].id");
   _.set(smartNotifierPayload, "connector.private_settings.synchronized_segments", [userSegmentId]);
 
-  _.set(smartNotifierPayload, "messages[0].user.traits_customerio/created_at", _.get(smartNotifierPayload, "messages[0].user.created_at"));
-  _.set(smartNotifierPayload, "messages[0].user.traits_customerio/email", _.get(smartNotifierPayload, "messages[0].user.email"));
-  _.set(smartNotifierPayload, "messages[0].user.traits_customerio/id", _.get(smartNotifierPayload, "messages[0].user.email"));
+  _.set(smartNotifierPayload, "messages[0].user.customerio/created_at", _.get(smartNotifierPayload, "messages[0].user.created_at"));
+  _.set(smartNotifierPayload, "messages[0].user.customerio/email", _.get(smartNotifierPayload, "messages[0].user.email"));
+  _.set(smartNotifierPayload, "messages[0].user.customerio/id", _.get(smartNotifierPayload, "messages[0].user.email"));
 
   const segmentIds = _.get(smartNotifierPayload, "messages[0].user.segment_ids", []);
 
@@ -24,7 +24,7 @@ module.exports = () => {
 
   const hashUtil = new HashUtil();
   const hash = hashUtil.hash(customerData);
-  _.set(smartNotifierPayload, "messages[0].user.traits_customerio/hash", hash);
+  _.set(smartNotifierPayload, "messages[0].user.customerio/hash", hash);
 
   return smartNotifierPayload;
 };